refactor: constraint submitter
