+++ crates/preconfer/src/lib.rs
@@ -1,4 +1,5 @@
 mod clients;
+mod constraint_submit;
 pub mod context_ext;
 mod contract;
 mod error;
