+++ crates/preconfer/src/preconf_api/state.rs
@@ -1,34 +1,21 @@
 use std::{
-    future::Future,
     sync::Arc,
-    time::{Duration, SystemTime, UNIX_EPOCH},
+    time::{SystemTime, UNIX_EPOCH},
 };
 
-use alloy_consensus::{Header, Transaction};
-use alloy_eips::{eip1559::BaseFeeParams, eip2718::Encodable2718, BlockId};
-use alloy_network::{EthereumWallet, TransactionBuilder};
-use alloy_primitives::{
-    keccak256, private::alloy_rlp::Decodable, Address, Bytes, PrimitiveSignature, U256,
-};
-use alloy_provider::{ext::DebugApi, utils::EIP1559_MIN_PRIORITY_FEE, Provider};
-use ethereum_consensus::{
-    clock::from_system_time, deneb::mainnet::MAX_BYTES_PER_TRANSACTION, primitives::BlsPublicKey,
-    ssz::prelude::ByteList,
-};
-use futures::StreamExt;
+use alloy_consensus::Transaction;
+use alloy_primitives::{Address, PrimitiveSignature};
+use alloy_provider::Provider;
 use reqwest::Url;
 use taiyi_primitives::{
-    BlockspaceAllocation, ConstraintsMessage, PreconfRequest, PreconfRequestTypeA,
-    PreconfRequestTypeB, PreconfResponse, SignableBLS, SignedConstraints, SlotInfo,
-    SubmitTransactionRequest, SubmitTypeATransactionRequest, TxExt,
+    BlockspaceAllocation, PreconfRequestTypeA, PreconfRequestTypeB, PreconfResponse, SlotInfo,
+    SubmitTransactionRequest, SubmitTypeATransactionRequest,
 };
-use tracing::{debug, error, info};
 use uuid::Uuid;
 
 use crate::{
     clients::{relay_client::RelayClient, signer_client::SignerClient},
     context_ext::ContextExt,
-    contract::{core::TaiyiCore, to_solidity_type},
     error::{PoolError, RpcError},
     network_state::NetworkState,
     preconf_pool::{PreconfPool, PreconfPoolBuilder},
@@ -36,12 +23,12 @@ use crate::{
 
 #[derive(Clone)]
 pub struct PreconfState<P> {
-    network_state: NetworkState,
-    preconf_pool: Arc<PreconfPool>,
-    relay_client: RelayClient,
-    signer_client: SignerClient,
-    provider: P,
-    min_fee_per_gas: u128,
+    pub network_state: NetworkState,
+    pub preconf_pool: Arc<PreconfPool>,
+    pub relay_client: RelayClient,
+    pub signer_client: SignerClient,
+    pub provider: P,
+    pub min_fee_per_gas: u128,
 }
 
 impl<P> PreconfState<P>
@@ -59,291 +46,6 @@ where
     ) -> Self {
         let preconf_pool = PreconfPoolBuilder::new().build(execution_rpc_url, taiyi_escrow_address);
         Self { relay_client, network_state, preconf_pool, signer_client, provider, min_fee_per_gas }
-    }
-
-    pub fn spawn_constraint_submitter(self) -> impl Future<Output = eyre::Result<()>> {
-        let relay_client = self.relay_client.clone();
-        let context = self.network_state.context();
-        let genesis_time = match context.genesis_time() {
-            Ok(genesis_time) => genesis_time,
-            Err(_) => context.min_genesis_time + context.genesis_delay,
-        };
-        let chain_id = self.network_state.chain_id();
-        info!("Starting constraint submitter, chain_id: {chain_id}");
-
-        async move {
-            let clock =
-                from_system_time(genesis_time, context.seconds_per_slot, context.slots_per_epoch);
-            let mut slot_stream = clock.into_stream();
-            while let Some(slot) = slot_stream.next().await {
-                let next_slot = slot + 1;
-
-                let submit_constraint_deadline_duration =
-                    context.get_deadline_of_slot(next_slot).saturating_sub(
-                        SystemTime::now()
-                            .duration_since(UNIX_EPOCH)
-                            .expect("Time went backwards")
-                            .as_secs(),
-                    );
-
-                // calculate base fee for next slot based on parent header
-                // Its fine to use latest block as we are submitting constraints for next block
-                let rlp_encoded_header =
-                    self.provider.debug_get_raw_header(BlockId::latest()).await?;
-                let header = Header::decode(&mut rlp_encoded_header.as_ref())?;
-                let (base_fee, priority_fee) =
-                    match header.next_block_base_fee(BaseFeeParams::ethereum()) {
-                        Some(base_fee) => (base_fee.into(), EIP1559_MIN_PRIORITY_FEE),
-                        None => {
-                            let estimate = self.provider.estimate_eip1559_fees(None).await?;
-                            (estimate.max_fee_per_gas, estimate.max_priority_fee_per_gas)
-                        }
-                    };
-                let blob_fee = header.next_block_blob_fee().unwrap_or_default();
-                let blob_excess_fee = header.next_block_excess_blob_gas().unwrap_or_default();
-
-                // wait unit the deadline to submit constraints
-                tokio::time::sleep(Duration::from_secs(submit_constraint_deadline_duration)).await;
-
-                info!(base_fee=?base_fee, priority_fee=?priority_fee, blob_fee=?blob_fee, blob_excess_fee=?blob_excess_fee);
-
-                let signer = self.signer_client.ecdsa_signer();
-                let wallet = EthereumWallet::from(signer.clone());
-                let sender = self.signer_client.ecdsa_address();
-
-                let taiyi_core =
-                    TaiyiCore::new(self.preconf_pool.taiyi_escrow_address, self.provider.clone());
-
-                let mut constraints = Vec::new();
-                let mut sponsoring_tx = Vec::new();
-                let mut type_a_txs = Vec::new();
-                let mut type_b_txs = Vec::new();
-                let mut exhaust_txs = Vec::new();
-
-                let mut nonce = self.provider.get_transaction_count(sender).await?;
-                // Accounts to sponsor gas for
-                let mut accounts = Vec::new();
-                // Amounts to sponsor for each account
-                let mut amounts = Vec::new();
-
-                match self.preconf_pool.ready_requests(next_slot) {
-                    Ok(preconf_requests) => {
-                        let sponsor_nonce = nonce;
-                        nonce += 1;
-                        for preconf_req in preconf_requests {
-                            match preconf_req {
-                                PreconfRequest::TypeA(request) => {
-                                    let tip_tx_gas_uesd = self
-                                        .preconf_pool
-                                        .calculate_gas_used(request.tip_transaction.clone())
-                                        .await?;
-                                    let mut preconf_tx_gas_used: u64 = 0;
-                                    for preconf_tx in request.preconf_tx.clone() {
-                                        let gas_used = self
-                                            .preconf_pool
-                                            .calculate_gas_used(preconf_tx)
-                                            .await?;
-                                        preconf_tx_gas_used += gas_used;
-                                    }
-
-                                    accounts
-                                        .push(request.signer().expect("Signer must be present"));
-                                    amounts.push(U256::from(
-                                        (tip_tx_gas_uesd + preconf_tx_gas_used) as u128 * base_fee,
-                                    ));
-
-                                    let tx_bytes = request.tip_transaction.to_ssz_bytes();
-                                    type_a_txs.push(tx_bytes);
-
-                                    for preconf_tx in request.preconf_tx {
-                                        let tx_bytes = preconf_tx.to_ssz_bytes();
-                                        type_a_txs.push(tx_bytes);
-                                    }
-                                }
-                                PreconfRequest::TypeB(preconf_req) => {
-                                    if let Some(ref tx) = preconf_req.transaction {
-                                        // calculate gas used
-                                        let gas_used = self
-                                            .preconf_pool
-                                            .calculate_gas_used(tx.clone())
-                                            .await?;
-
-                                        accounts.push(
-                                            preconf_req.signer().expect("Signer must be present"),
-                                        );
-                                        amounts.push(U256::from(gas_used as u128 * base_fee));
-
-                                        // preconf tx
-                                        let mut tx_encoded = Vec::new();
-                                        tx.encode_2718(&mut tx_encoded);
-                                        let tx_ref: &[u8] = tx_encoded.as_ref();
-                                        let tx_bytes: ByteList<MAX_BYTES_PER_TRANSACTION> =
-                                            tx_ref.try_into().expect("tx bytes too big");
-                                        type_b_txs.push(tx_bytes);
-
-                                        // Append with a transaction that calls get_tip() on TaiyiCore contract
-                                        let blockspace_allocation_sig_user = preconf_req.alloc_sig;
-                                        let blockspace_allocation_sig_gateway = self
-                                            .signer_client
-                                            .sign_with_ecdsa(preconf_req.allocation.digest())
-                                            .await
-                                            .map_err(|e| {
-                                                RpcError::SignatureError(format!("{e:?}"))
-                                            })?;
-                                        let gateway_signed_raw_tx = self
-                                            .signer_client
-                                            .sign_with_ecdsa(keccak256(tx_encoded.clone()))
-                                            .await
-                                            .map_err(|e| {
-                                                RpcError::SignatureError(format!(
-                                                    "Failed to issue commitment: {e:?}"
-                                                ))
-                                            })?;
-                                        let preconf_request_type_b = to_solidity_type(
-                                            preconf_req,
-                                            blockspace_allocation_sig_user,
-                                            blockspace_allocation_sig_gateway,
-                                            tx_encoded.into(),
-                                            gateway_signed_raw_tx,
-                                            self.preconf_pool.taiyi_escrow_address,
-                                        );
-
-                                        // Call getTip() on TaiyiCore contract
-                                        let get_tip_tx = taiyi_core
-                                            .getTip(preconf_request_type_b)
-                                            .into_transaction_request()
-                                            .with_chain_id(chain_id)
-                                            .with_nonce(nonce)
-                                            .with_gas_limit(1_000_000)
-                                            .with_max_fee_per_gas(base_fee)
-                                            .with_max_priority_fee_per_gas(priority_fee)
-                                            .build(&wallet)
-                                            .await?;
-                                        // increment nonce
-                                        nonce += 1;
-                                        let tx_bytes = get_tip_tx.to_ssz_bytes();
-                                        type_b_txs.push(tx_bytes);
-                                    }
-                                }
-                            }
-                        }
-
-                        //  gas sponsorship tx
-                        let sponsor_tx = taiyi_core
-                            .sponsorEthBatch(accounts, amounts)
-                            .into_transaction_request()
-                            .with_nonce(sponsor_nonce)
-                            .with_chain_id(chain_id)
-                            .with_gas_limit(1_000_000)
-                            .with_max_fee_per_gas(base_fee)
-                            .with_max_priority_fee_per_gas(priority_fee)
-                            .build(&wallet)
-                            .await?;
-
-                        let tx_bytes = sponsor_tx.to_ssz_bytes();
-                        sponsoring_tx.push(tx_bytes);
-                    }
-                    Err(err) => {
-                        debug!(?err, "Error fetching preconf requests for slot");
-                    }
-                }
-
-                // Fetch all preconf requests for which the gateway must call exhaust() on TaiyiCore contract
-                let requests = self.preconf_pool.fetch_pending(next_slot);
-                if let Some(requests) = requests {
-                    info!(
-                        "Found {} preconf requests for slot {} to be exhausted",
-                        requests.len(),
-                        next_slot
-                    );
-
-                    for preconf_req in requests {
-                        let blockspace_allocation_sig_user = preconf_req.alloc_sig;
-                        let blockspace_allocation_sig_gateway = self
-                            .signer_client
-                            .sign_with_ecdsa(preconf_req.allocation.digest())
-                            .await
-                            .map_err(|e| {
-                                RpcError::SignatureError(format!(
-                                    "Failed to issue commitment: {e:?}"
-                                ))
-                            })?;
-                        let preconf_request_type_b = to_solidity_type(
-                            preconf_req,
-                            blockspace_allocation_sig_user,
-                            blockspace_allocation_sig_gateway,
-                            Bytes::default(),
-                            self.signer_client
-                                .sign_with_ecdsa(keccak256(Bytes::default()))
-                                .await
-                                .map_err(|e| {
-                                RpcError::SignatureError(format!(
-                                    "Failed to issue commitment: {e:?}"
-                                ))
-                            })?,
-                            self.preconf_pool.taiyi_escrow_address,
-                        );
-
-                        // Call exhaust() on TaiyiCore contract
-                        let exhaust_tx = taiyi_core
-                            .exhaust(preconf_request_type_b)
-                            .into_transaction_request()
-                            .with_chain_id(chain_id)
-                            .with_nonce(nonce)
-                            .with_gas_limit(1_000_000)
-                            .with_max_fee_per_gas(base_fee)
-                            .with_max_priority_fee_per_gas(priority_fee)
-                            .build(&wallet)
-                            .await?;
-                        // increment nonce
-                        nonce += 1;
-
-                        let tx_bytes = exhaust_tx.to_ssz_bytes();
-                        exhaust_txs.push(tx_bytes);
-                    }
-                }
-
-                constraints.extend(sponsoring_tx);
-                constraints.extend(type_a_txs);
-                constraints.extend(type_b_txs);
-                constraints.extend(exhaust_txs);
-
-                let txs_len = constraints.len();
-                if txs_len != 0 {
-                    let bls_pk = self.signer_client.bls_pubkey();
-                    let message = ConstraintsMessage {
-                        pubkey: BlsPublicKey::try_from(bls_pk.to_bytes().as_ref())
-                            .expect("key error"),
-                        slot: next_slot,
-                        top: false,
-                        transactions: constraints.try_into().expect("tx too big"),
-                    };
-                    let digest = message.digest();
-                    if let Ok(signature) = self.signer_client.sign_with_bls(context.clone(), digest)
-                    {
-                        let signed_constraints_message =
-                            vec![SignedConstraints { message, signature }];
-
-                        let max_retries = 5;
-                        let mut i = 0;
-
-                        info!("Submitting {txs_len} constraints to relay on  slot {next_slot}");
-                        'submit: while let Err(e) =
-                            relay_client.set_constraints(signed_constraints_message.clone()).await
-                        {
-                            error!(err = ?e, "Error submitting constraints to relay, retrying...");
-                            i += 1;
-                            if i >= max_retries {
-                                error!("Max retries reached while submitting to relay");
-                                break 'submit;
-                            }
-                            tokio::time::sleep(Duration::from_millis(100)).await;
-                        }
-                    }
-                }
-            }
-            Ok(())
-        }
     }
 
     /// reserve blockspace for a slot
